+++ src/main/nextflow/co2footprint/CO2FootprintPlugin.groovy
@@ -81,7 +81,8 @@ class CO2FootprintPlugin extends BasePlugin implements PluginAbstractExec {
                 [observer.workflowStats]
         )
 
-        observer.renderFiles()
+        // Render files
+        observer.renderFiles(sessionStats)
 
         super.stop()
     }